+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/TopShops2.iml" filepath="$PROJECT_DIR$/TopShops2.iml" />
+      <module fileurl="file://$PROJECT_DIR$/TopShops.iml" filepath="$PROJECT_DIR$/TopShops.iml" />
     </modules>
   </component>
 </project>
