--- src/product.java
@@ -1,28 +1,0 @@
-public class product {
-    private String name, price, company, type, event, totalSales, salesBy7;
-
-    public product() {}
-
-    //getters
-    public String getName() {return name;}
-    public String getPrice() {return price;}
-    public String getCompany() {return company;}
-    public String getType() {return type;}
-    public String getEvent() {return event;}
-    public String getTotalSales() {return totalSales;}
-    public String getSalesBy7() {return salesBy7;}
-
-    //setters
-    public void setName(String name) {this.name = name;}
-    public void setPrice(String price) {this.price = price;}
-    public void setCompany(String company) {this.company = company;}
-    public void setType(String type) {this.type = type;}
-    public void setEvent(String event) {this.event = event;}
-    public void setTotalSales(String totalSales) {this.totalSales = totalSales;}
-    public void setSalesBy7(String salesBy7) {this.salesBy7 = salesBy7;}
-
-    //
-    public void printProducts() {
-
-    }
-}
